msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

edit: adapted for op7/msm4.14

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: grislux55 <jebchou.hg@gmail.com>