usb: dwc3: gadget: Ensure minimum delay between run/stop
Some controller hardware, such as DWC3, need a minimum delay
between removing and re-applying the pullups in order for the
host to properly detect a soft disconnect and for subsequent
re-enumeration to succeed. Keep track of the last time run_stop
was updated and delay if necessary so that enough time will
have elapsed before calling run_stop clear/set again.

Change-Id: I8ca13304edf4c9fcc3093a063a6004bd8b76ed44
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>